Modified initialisation i

+++ singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t num = 0;
+	int num = 0;
 	const list_t *current = h;
 
 	while (current != NULL)
